+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 ## Versions
 
   2.3.5
+    * Sorting on last name in agenda, message receivers and groups&users listings 26-8-2013
+    * Download exp. mobile app 23-8-2013
 
   2.3.4
 	  * Deployed for 100T, 110K, 120G and 130B 21-8-2013